refactor(#3037): relocate round-table meeting DTOs to services (backflow 22→21)
Move the pure-serde round-table persistence request DTOs `UpsertMeetingBody`
and `MeetingEntryBody` from `crate::server::routes::meetings` down to the
service-layer meeting module (`services::discord::meeting_artifact_store`),
beside their producers (`meeting_orchestrator`, `internal_api`). The server
route now `use`s `UpsertMeetingBody` from services instead of owning it.

These structs have no axum / `AppState` coupling, so the move is a pure
relocation — identical fields and derives, behavior unchanged. It removes the
`meeting_orchestrator.rs` service→server backflow reference and touches no hot
files (turn_bridge/completion_guard, tmux_watcher, etc.).

- Cluster: round-table meeting persistence DTOs (server route → services).
- backflow 22→21 (meeting_orchestrator.rs entry removed from baseline).
- Behavior invariant; no circular dependency (services owns the types,
  server route depends on services — the correct direction).
- All call sites updated: internal_api, meetings route handler, orchestrator
  persist path.
- Gates: cargo fmt --check clean; cargo check --lib / --lib --tests clean
  (0 new warnings); escalation pg test green; audit_maintainability --check
  exit 0 (backflow 21, giant-file ratchet 0); inventory + maintenance-docs
  freshness passed (meetings.rs change-surfaces freeze synced 1708→1686);
  ci-script-checks exit 0.

No multinode-audited surface touched (no ws/cluster/AppState changes).

Refs #3037
